# Revert "Try ensuring the item after post content clears floats (#36006)"