--- a/packages/block-library/src/post-content/style.scss
+++ b/packages/block-library/src/post-content/style.scss
@@ -1,4 +1,0 @@
-// Required to ensure that subsequent blocks clear floats inside entry content.
-.wp-block-post-content + * {
-	clear: both;
-}