--- a/packages/block-library/src/style.scss
+++ b/packages/block-library/src/style.scss
@@ -31,7 +31,6 @@
 @import "./post-title/style.scss";
 @import "./preformatted/style.scss";
 @import "./pullquote/style.scss";
-@import "./post-content/style.scss";
 @import "./post-template/style.scss";
 @import "./query-pagination/style.scss";
 @import "./quote/style.scss";