While checking https://codehealth.vercel.app I noticed that the #36006 PR introduced significant regressions in both "block select" and "inserter open" metrics. I'm attempt a revert here to confirm the numbers.