@youknowriad 2021-11-08
I'm hesitant about merging this PR. The change in the "inserterOpen" metric should definitely be noticeable when testing manually but I'm unable to feel any difference personally. I think this might be a "metric" issue that is highlighted by this particular change but I'm not sure. Would love some thought if folks have ideas why this simple change seem to have such a big impact on the "inserterOpen" metric. cc @gwwar @talldan @ockham