@gwwar 2021-11-08
Interesting find @youknowriad.

For folks who can't see the performance test runs, it looks like changes here affect inserter open and list view open (basically anything that shifts the main block list layout).

<details>

```
>> post-editor

┌──────────────────────┬──────────────────────────────────────────┬──────────────┐
│       (index)        │ d421b429819963ad45c640fb60f4922c15129d84 │    trunk     │
├──────────────────────┼──────────────────────────────────────────┼──────────────┤
│    serverResponse    │               '245.84 ms'                │  '240.2 ms'  │
│      firstPaint      │                '47.66 ms'                │  '56.84 ms'  │
│   domContentLoaded   │               '329.28 ms'                │ '352.78 ms'  │
│        loaded        │               '339.66 ms'                │  '363.9 ms'  │
│ firstContentfulPaint │               '6121.84 ms'               │ '6157.1 ms'  │
│      firstBlock      │               '7097.9 ms'                │ '7005.94 ms' │
│         type         │                '41.89 ms'                │  '38.57 ms'  │
│       minType        │                '33.08 ms'                │  '32.29 ms'  │
│       maxType        │                '55.36 ms'                │  '52.91 ms'  │
│        focus         │               '159.34 ms'                │ '185.37 ms'  │
│       minFocus       │               '125.32 ms'                │ '120.21 ms'  │
│       maxFocus       │               '221.46 ms'                │ '515.32 ms'  │
│     inserterOpen     │                '92.33 ms'                │ '381.72 ms'  │
│   minInserterOpen    │                '58.6 ms'                 │ '330.04 ms'  │
│   maxInserterOpen    │                '304.2 ms'                │ '645.27 ms'  │
│    inserterSearch    │                '94.66 ms'                │  '88.43 ms'  │
│  minInserterSearch   │                '70.02 ms'                │  '68.81 ms'  │
│  maxInserterSearch   │               '121.07 ms'                │ '118.56 ms'  │
│    inserterHover     │                '35.39 ms'                │  '36.43 ms'  │
│   minInserterHover   │                '29.04 ms'                │  '29.38 ms'  │
│   maxInserterHover   │                '45.96 ms'                │  '49.27 ms'  │
│     listViewOpen     │               '287.97 ms'                │ '601.92 ms'  │
│   minListViewOpen    │               '254.38 ms'                │ '563.46 ms'  │
│   maxListViewOpen    │               '353.11 ms'                │ '658.42 ms'  │
└──────────────────────┴──────────────────────────────────────────┴──────────────┘

>> site-editor

┌──────────────────────┬──────────────────────────────────────────┬──────────────┐
│       (index)        │ d421b429819963ad45c640fb60f4922c15129d84 │    trunk     │
├──────────────────────┼──────────────────────────────────────────┼──────────────┤
│    serverResponse    │               '166.27 ms'                │ '154.07 ms'  │
│      firstPaint      │               '641.03 ms'                │ '511.13 ms'  │
│   domContentLoaded   │               '693.27 ms'                │ '560.07 ms'  │
│        loaded        │               '749.53 ms'                │ '609.87 ms'  │
│ firstContentfulPaint │               '641.03 ms'                │ '511.13 ms'  │
│      firstBlock      │               '7935.1 ms'                │ '7784.37 ms' │
│         type         │                '36.26 ms'                │  '34.8 ms'   │
│       minType        │                '26.9 ms'                 │  '26.41 ms'  │
│       maxType        │                '80.73 ms'                │  '67.1 ms'   │
└──────────────────────┴──────────────────────────────────────────┴──────────────┘
```
</details>

I profiled and it does look like there is a difference for how much time is spent in recalculating styles. On my machine the difference was more around 50ms with a post of 900 blocks but it may be more pronounced on the test profiles we setup. I would recommend reverting this and finding an alternative.

To reproduce, start the profiler, click the inserter open/closed **twice**. (Note that the code hotspot times are accumulative)> in the click event inspect the purple style recalculation events.

| before | before hotspot | after | after hotspot |
|-----|-----|-----|-----|
| ![CleanShot 2021-11-08 at 10 45 20](https://user-images.githubusercontent.com/1270189/140800277-7f9e83a6-9d8e-4233-8fdb-c009c97aa9e8.png) |  ![CleanShot 2021-11-08 at 10 44 55](https://user-images.githubusercontent.com/1270189/140800312-45cd76f4-e62e-4241-a8b4-9738d027e86a.png) | ![CleanShot 2021-11-08 at 10 47 13](https://user-images.githubusercontent.com/1270189/140800358-cb5d7ca7-316a-46ac-9ee8-3ed8018a9c04.png) | ![CleanShot 2021-11-08 at 10 47 02](https://user-images.githubusercontent.com/1270189/140800435-71e399ee-fc47-48d2-9a7c-fd5ce9cc3ea9.png) |

Could we maybe add a note to the next performance meeting to help highlight perf jumps/improvements on the PR? I don't think we need a blocking check but it'd be ideal to have a visualization, like we do for package-size changes.