@gwwar 2021-11-08
As a follow up note, the style recalculation is pretty sensitive to CPU speeds. Testing trunk again on my machine with CPU x6 slowdown, we see that time spent goes from ~50ms to ~240ms

![CleanShot 2021-11-08 at 11 14 39](https://user-images.githubusercontent.com/1270189/140804155-30728502-1d0f-4645-9495-04df550df943.png)