@youknowriad 2021-11-09
Thanks for the analysis @gwwar let's merge this then and see if we can find an alternative.