@andrewserong 2021-11-10
Cherry picked into the Gutenberg 11.9 release in: https://github.com/WordPress/gutenberg/commit/d6f4f80d3eba3433489f546004e0e1a6b775c4bc